refactor: 자동차 이름 프로퍼티를 CarName 클래스로 분리
