+++ src/main/kotlin/racingcar/view/OutputView.kt
@@ -1,19 +1,19 @@
 package racingcar.view
 
-import racingcar.model.Car
+import racingcar.model.car.Car
 
 class OutputView {
     fun printMessage(message: String) = println(message)
 
     fun printRoundResult(cars: List<Car>) {
         cars.forEach { car ->
-            println("${car.name} : ${getPositionAsDash(car.position)}")
+            println("${car.carName.name} : ${getPositionAsDash(car.position)}")
         }
         println()
     }
 
     fun printWinners(winners: List<Car>) {
-        println("$WINNER_NOTIFICATION_MESSAGE: ${winners.joinToString(", ") { it.name }}")
+        println("$WINNER_NOTIFICATION_MESSAGE: ${winners.joinToString(", ") { it.carName.name }}")
     }
 
     private fun getPositionAsDash(position: Int): String = DASH.repeat(position)
