+++ src/test/kotlin/racingcar/model/CarTest.kt
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.assertDoesNotThrow
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
+import racingcar.model.car.Car
 import kotlin.test.assertNotEquals
 
 class CarTest {
